pyproject.toml
[tool.poetry]
name = "collect_nasa_lis_data"
version = "0.1.0"
description = ""
authors = ["Joshua Maldonado <joshjm9915@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.10.0"
requests = "^2.27.1"
netCDF4 = "^1.5.8"
numpy = "^1.22.1"
pandas = "^1.4.0"
scikit-learn = "^1.0.2"
matplotlib = "^3.5.1"
geopy = "^2.2.0"
Shapely = "^1.8.0"
jupyterlab = "^3.2.8"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
